pyproject.toml
[project]
name = "resume-builder-drf"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "redis (>=7.0.1,<8.0.0)",
]

[tool.poetry]
packages = [{ include = "resume_builder_drf", from = "src" }]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.black]
line-length = 88

[tool.django-stubs]
django_settings_module = "myproject.settings"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mypy (>=1.18.2,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "django-stubs (>=5.2.7,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
]
